Added check for available service at startup
App now responds to situations where a connection to a network cannot be made or geolocation fails. A modal pops up telling the user to make sure a connection is present and to ensure location services are on. User can now close window by pressing the "Okay" button in the modal too.

Made a few tweaks to the U.I., nothing too big though. Finally added tabs to the bottom of the screen and moved the switch unit button to the right of those. App is now ready to have a Map and Search function added to it.

Porting to mobile will be an interesting experience. Currently Chrome Dev Tools doesn't like the app, but I'll see what I can do to make it work.