Started adding layers to map
Was able to get precipitation and clouds to show up, although it is slow because it grabs it from OWM's tile server. Clouds are a bit buggy though, so it's layer is currently disabled. Some how broke animations, not sure how, code is identical to the previous commit. That will be fixed later, I'm just happy the map actually works.